+++ cdl/tests/features/signal/operation_unit_test.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-import cdl.core.computation.signal as cps
+import cdl.computation.signal as cps
 import cdl.obj
 import cdl.param
 import cdl.tests.data as ctd